+++ theme/template-parts/entry-list-item.php
@@ -7,6 +7,7 @@
 
 declare(strict_types=1);
 
+use DiarioDelNorte\Content\SubscriberOnly;
 use DiarioDelNorte\Support\Format;
 
 if ( ! defined( 'ABSPATH' ) ) {
@@ -24,6 +25,9 @@ $ddn_cat = Format::primary_category();
 	<div class="entry-row__body">
 		<h3 class="entry-row__title">
 			<a class="headline-link" href="<?php the_permalink(); ?>"><?php the_title(); ?></a>
+			<?php if ( SubscriberOnly::is_restricted( get_the_ID() ) ) : ?>
+				<?php echo SubscriberOnly::badge_markup(); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped -- ya escapado en badge_markup(). ?>
+			<?php endif; ?>
 		</h3>
 		<span class="meta">
 			<?php